requirements.txt
aiogram==3.10.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
apscheduler==3.10.4
requests==2.32.4
aiohttp==3.12.14